requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0

# LLM
anthropic==0.28.0

# Embeddings & Vector Store
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4

# Document Processing
pypdf==4.2.0
python-docx==1.1.2
unstructured==0.14.4
langchain-text-splitters==0.2.1

# Observability & Governance
structlog==24.2.0
python-json-logger==2.0.7

# Dev / Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
tiktoken==0.7.0
